+++ for-vue2/src/packages/utils/Hhttp/Hhttp.js
@@ -36,6 +36,7 @@ export default class Hhttp {
 
   // 请求配置策略
   static getUrl = (baseUrl, url = '') => {
+    if (!baseUrl) return url;
     const base = baseUrl.replace(/\/+$/, ''); // 去掉baseUrl末尾的斜杠
     const path = url.replace(/^\/+/, ''); // 去掉url开头的斜杠
     return `${base}/${path}`;