+++ for-vue2/src/pages/HTab/HTabDirection.vue
@@ -25,11 +25,11 @@
       :direction="tabDirection"
     >
       <HTabItem
-        v-for="( item, index ) in 20"
+        v-for="( item, index ) in [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]"
         ref="tabItem"
         :key="index"
         :direction="tabItemDirection"
-        :value="item"
+        :value="index"
         :label="`套餐${item}`"
         image="/static/logo-shadow.png"
       />
@@ -44,7 +44,7 @@ export default {
   data() {
     return {
 
-      active: 1,
+      active: 0,
       tabDirection: 'y',
       tabItemDirection: 'y',