+++ for-vue2/src/pages/HTab/HTabItemStyle.vue
@@ -1,6 +1,6 @@
 <!-- #region template -->
 <template>
-  <div>
+  <view>
     <uni-section
       title="选项卡样式"
       type="line"
@@ -11,7 +11,7 @@
         direction="x"
       >
         <HTabItem
-          v-for="(item, index) in 20"
+          v-for="(item, index) in [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]"
           :key="index"
           :value="index"
           :label="`选项${item}`"
@@ -34,7 +34,7 @@
       >
         <HTab v-model="active1">
           <HTabItem
-            v-for="(item, index) in 20"
+            v-for="(item, index) in [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]"
             :key="index"
             :value="index"
             :label="`选项${item}`"
@@ -56,7 +56,7 @@
       >
         <HTab v-model="active2">
           <HTabItem
-            v-for="(item, index) in 20"
+            v-for="(item, index) in [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]"
             :key="index"
             :value="index"
             :label="`选项${item}`"
@@ -78,7 +78,7 @@
       >
         <HTab v-model="active3">
           <HTabItem
-            v-for="(item, index) in 20"
+            v-for="(item, index) in [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]"
             :key="index"
             :value="index"
             :label="`选项${item}`"
@@ -94,7 +94,7 @@
       >
         <HTab v-model="active4">
           <HTabItem
-            v-for="(item, index) in 20"
+            v-for="(item, index) in [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19, 20]"
             :key="index"
             :value="index"
             :label="`选项${item}`"
@@ -104,7 +104,7 @@
         </HTab>
       </uni-section>
     </view>
-  </div>
+  </view>
 </template>
 <!-- #endregion template -->