+++ for-vue2/src/pages/TestCom.vue
@@ -1,0 +1,18 @@
+<template>
+  <view>
+    {{ num }}
+  </view>
+</template>
+
+<script>
+export default {
+  props: {
+    num: {
+      type: Number,
+      default: 0,
+    },
+  },
+  methods: {
+  },
+};
+</script>